pre submission

+++ Inception/srcs/requirements/wordpress/tools/config.sh
@@ -35,15 +35,20 @@ fi
 mkdir -p var/www/html/wordpress /run/php/
 cd /var/www/html/
 
-# WordPress setup commands
-wp core download --allow-root
-wp config create --dbname=$DB_NAME --dbuser=$DB_USER --dbpass=$DB_USER_PASSWORD --dbhost=mariadb --path=/var/www/html/ --force
-wp core install --url=$DOMAIN_NAME --title=$WP_TITLE --admin_user=$WP_ADMIN_USER --admin_password=$WP_ADMIN_PASSWORD --admin_email=$WP_ADMIN_EMAIL --skip-email --path=/var/www/html/ --allow-root
-wp user create $WP_USER $WP_USER_EMAIL --role=editor --user_pass=$WP_USER_PASSWORD --path=/var/www/html/
-wp theme install twentysixteen --activate --allow-root
-wp plugin update --all
-wp option update siteurl "https://$DOMAIN_NAME"
-wp option update home "https://$DOMAIN_NAME"
+if [ -f "./.wpExists" ]; then
+    echo "Wordpress already exists"
+else
+	touch ./.wpExists
+	# WordPress setup commands
+	wp core download --allow-root
+	wp config create --dbname=$DB_NAME --dbuser=$DB_USER --dbpass=$DB_USER_PASSWORD --dbhost=mariadb --path=/var/www/html/ --force
+	wp core install --url=$DOMAIN_NAME --title=$WP_TITLE --admin_user=$WP_ADMIN_USER --admin_password=$WP_ADMIN_PASSWORD --admin_email=$WP_ADMIN_EMAIL --skip-email --path=/var/www/html/ --allow-root
+	wp user create $WP_USER $WP_USER_EMAIL --role=editor --user_pass=$WP_USER_PASSWORD --path=/var/www/html/
+	wp theme install twentysixteen --activate --allow-root
+	wp plugin update --all
+	wp option update siteurl "https://$DOMAIN_NAME"
+	wp option update home "https://$DOMAIN_NAME"
+fi
 
 # Change ownership and permissions
 chown -R www:www /var/www/html
